requirements.txt
streamlit
groq
pillow
gtts
requests